requirements.txt
beautifulsoup4
boto3
botocore
bs4
lxml
numpy
pandas
python-dotenv
requests
selenium
slackweb